refactor: 불필요한 코드 정리 및 파일 위치 이동
