+++ src/main/java/org/server/remoteclass/controller/OrderController.java
@@ -1,8 +1,8 @@
 package org.server.remoteclass.controller;
 
 import io.swagger.annotations.ApiOperation;
-import org.server.remoteclass.dto.*;
-import org.server.remoteclass.entity.OrderLecture;
+import org.server.remoteclass.dto.order.RequestOrderDto;
+import org.server.remoteclass.dto.order.ResponseOrderDto;
 import org.server.remoteclass.exception.ForbiddenException;
 import org.server.remoteclass.exception.IdNotExistException;
 import org.server.remoteclass.service.OrderService;
