+++ src/main/java/org/server/remoteclass/dto/order/OrderDto.java
@@ -1,4 +1,4 @@
-package org.server.remoteclass.dto;
+package org.server.remoteclass.dto.order;
 
 import lombok.*;
 import org.server.remoteclass.constant.OrderStatus;
@@ -10,6 +10,7 @@ import org.server.remoteclass.entity.User;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
+import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -19,11 +20,14 @@ import java.util.List;
 @Setter
 public class OrderDto {
     private Long orderId;
+    @NotNull
     private Long userId;
+    @NotNull
     private List<OrderLecture> orderLectures = new ArrayList<>();
     private OrderStatus orderStatus; //주문상태
     private LocalDateTime orderDate;
     private Long couponId;       //적용하는 쿠폰 아이디
+    @NotNull
     private Payment payment; //결제방법
     private String bank;  //입금은행
     private String account;  //예금주
