+++ src/main/java/org/server/remoteclass/dto/order/ResponseOrderDto.java
@@ -1,34 +1,38 @@
-package org.server.remoteclass.dto;
+package org.server.remoteclass.dto.order;
 
 
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.server.remoteclass.constant.OrderStatus;
 import org.server.remoteclass.constant.Payment;
-import org.server.remoteclass.entity.*;
+import org.server.remoteclass.entity.Order;
 
-
-import java.time.LocalDateTime;
+import javax.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Builder
-@Getter
-@Setter
-public class RequestOrderDto {
+@Getter @Setter
+@NoArgsConstructor @AllArgsConstructor
+public class ResponseOrderDto {
 
+    private Long orderId;
+    @NotNull
+    private Long userId;
+    @NotNull
     private List<ResponseOrderLectureDto> orderLectures = new ArrayList<>();
     private OrderStatus orderStatus; //주문상태
+    @NotNull
     private Payment payment; //결제방법
     private String bank;  //입금은행
     private String account;  //예금주
     private Long couponId;       //적용하는 쿠폰 아이디
 
-    public static RequestOrderDto from(Order order){
+    public static ResponseOrderDto from(Order order){
         if(order == null) return null;
-        return RequestOrderDto.builder()
+        return ResponseOrderDto.builder()
+                .userId(order.getUser().getUserId())
+                .orderId(order.getOrderId())
                 .orderLectures(order.getOrderLectures().stream()
                         .map(ResponseOrderLectureDto::new)
                         .collect(Collectors.toList())
