+++ src/main/java/org/server/remoteclass/service/OrderService.java
@@ -1,7 +1,7 @@
 package org.server.remoteclass.service;
 
-import org.server.remoteclass.dto.RequestOrderDto;
-import org.server.remoteclass.dto.ResponseOrderDto;
+import org.server.remoteclass.dto.order.RequestOrderDto;
+import org.server.remoteclass.dto.order.ResponseOrderDto;
 import org.server.remoteclass.exception.ForbiddenException;
 import org.server.remoteclass.exception.IdNotExistException;
