+++ src/main/java/org/server/remoteclass/service/OrderServiceImpl.java
@@ -1,11 +1,10 @@
 package org.server.remoteclass.service;
 
-import io.swagger.annotations.ApiOperation;
 import org.modelmapper.ModelMapper;
-import org.server.remoteclass.constant.Authority;
 import org.server.remoteclass.constant.OrderStatus;
 import org.server.remoteclass.constant.Payment;
-import org.server.remoteclass.dto.*;
+import org.server.remoteclass.dto.order.RequestOrderDto;
+import org.server.remoteclass.dto.order.ResponseOrderDto;
 import org.server.remoteclass.entity.*;
 import org.server.remoteclass.exception.ForbiddenException;
 import org.server.remoteclass.exception.IdNotExistException;
@@ -19,7 +18,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import javax.persistence.EntityNotFoundException;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
